Move currencies sync-schema to separate file
