+++ workers/loc.api/sync/schema/sync-schema/index.js
@@ -30,6 +30,7 @@ const mapSymbols = require('./map-symbols')
 const inactiveCurrencies = require('./inactive-currencies')
 const inactiveSymbols = require('./inactive-symbols')
 const futures = require('./futures')
+const currencies = require('./currencies')
 
 const _methodCollMap = new Map([
   [SYNC_API_METHODS.LEDGERS, ledgers],
@@ -54,19 +55,7 @@ const _methodCollMap = new Map([
   [SYNC_API_METHODS.INACTIVE_CURRENCIES, inactiveCurrencies],
   [SYNC_API_METHODS.INACTIVE_SYMBOLS, inactiveSymbols],
   [SYNC_API_METHODS.FUTURES, futures],
-  [
-    SYNC_API_METHODS.CURRENCIES,
-    {
-      name: ALLOWED_COLLS.CURRENCIES,
-      maxLimit: 10000,
-      projection: null,
-      sort: [['name', 1]],
-      hasNewData: false,
-      isSyncRequiredAtLeastOnce: true,
-      type: COLLS_TYPES.PUBLIC_UPDATABLE_ARRAY_OBJECTS,
-      model: getModelOf(TABLES_NAMES.CURRENCIES)
-    }
-  ],
+  [SYNC_API_METHODS.CURRENCIES, currencies],
   [
     SYNC_API_METHODS.MARGIN_CURRENCY_LIST,
     {
